requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-decouple==3.6
requests==2.27.1
urllib3==1.26.8
wincertstore==0.2
wrapt==1.13.3
beautifulsoup4~=4.10.0
pandas~=1.4.1
termcolor~=1.1.0
numpy~=1.21.5
matplotlib~=3.5.1
jinja2~=3.0.2